+++ frontend/src/app/document-detail-page/document-detail-page.component.ts
@@ -1,20 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-
-import { AuthenticationService } from '../service/authentication.service';
-import { AlertService } from '../service/alert.service';
-import { BoardService } from '../service/board.service';
-import { BubbleService } from '../service/bubble.service';
-import { CommentService } from '../service/comment.service';
-import { DirectoryService } from '../service/directory.service';
-import { DocumentService } from '../service/document.service';
-import { NoteService } from '../service/note.service';
-
-import { Alert } from '../model/alert';
-import { Board } from '../model/board';
-import { BubbleType, Bubble, SuggestBubble } from '../model/bubble';
-import { Comment } from '../model/comment';
-import { FileSystemEntity } from '../model/file-system-entity';
-import { Note } from '../model/note';
+import { DocumentService } from './service';
 
 @Component({
   selector: 'app-document-detail-page',
@@ -59,6 +44,3 @@ export class DocumentDetailPageComponent implements OnInit {
     );
   }
 } /* istanbul ignore next */
-
-export { AuthenticationService, BoardService, BubbleService, AlertService, CommentService, DirectoryService, DocumentService, NoteService};
-export { Alert, Board, BubbleType, Bubble, SuggestBubble, Comment, FileSystemEntity, Note };
