+++ frontend/src/app/document-detail-page/edit-board/service.ts
@@ -1,0 +1,2 @@
+export { Note, NoteService } from '../service';
+
