+++ frontend/src/app/document-detail-page/service.ts
@@ -1,0 +1,15 @@
+export { AuthenticationService } from '../service/authentication.service';
+export { AlertService } from '../service/alert.service';
+export { BoardService } from '../service/board.service';
+export { BubbleService } from '../service/bubble.service';
+export { CommentService } from '../service/comment.service';
+export { DirectoryService } from '../service/directory.service';
+export { DocumentService } from '../service/document.service';
+export { NoteService } from '../service/note.service';
+
+export { Alert } from '../model/alert';
+export { Board } from '../model/board';
+export { BubbleType, Bubble, SuggestBubble } from '../model/bubble';
+export { Comment } from '../model/comment';
+export { FileSystemEntity } from '../model/file-system-entity';
+export { Note } from '../model/note';
