+++ frontend/src/app/document-detail-page/view-board/bubble-list-view/bubble-list-view.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed, async, fakeAsync, tick } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { BubbleService } from '../view-board.component';
-import { Bubble, MenuType } from '../view-board.component';
+import { BubbleService } from '../service';
+import { Bubble, MenuType } from '../service';
 import { BubbleListViewComponent } from './bubble-list-view.component';
 import { Component, Input } from '@angular/core';
 import { LeafBubble, InternalBubble } from '../../../model/bubble';
